Check for alluvial structure and convert between alluvial formats

Alluvial diagrams consist of multiple horizontally-distributed columns (axes) representing factor variables, vertical divisions (strata) of these axes representing these variables' values; and splines (alluvial flows) connecting vertical subdivisions (lodes) within strata of adjacent axes representing subsets or amounts of observations that take the corresponding values of the corresponding variables. This function checks a data frame for either of two types of alluvial structure:

One row per lode, wherein each row encodes a subset or amount of observations having a specific profile of axis values, a key field encodes the axis, a value field encodes the value within each axis, and a id column identifies multiple lodes corresponding to the same subset or amount of observations.

One row per alluvium, wherein each row encodes a subset or amount of observations having a specific profile of axis values and a set axes of fields encodes its values at each axis variable.

If no arguments are assigned to any of these parameters, then is_alluvial will default to is_alluvial_alluvia and assume that all fields in data (other than weight, if given) are to be treated as axes.

is_alluvial(data, ..., logical = TRUE, silent = FALSE)

is_alluvial_lodes(data, key, value, id, weight = NULL, logical = TRUE,
  silent = FALSE)

is_alluvial_alluvia(data, axes = NULL, weight = NULL, logical = TRUE,
  silent = FALSE)

to_lodes(data, key = "x", value = "stratum", id = "alluvium", axes,
  keep = FALSE)

to_alluvia(data, key, value, id, distill = FALSE)

Arguments

data
A data frame.

...
Additional parameters used to determine method and passed thereto. All or none of key, value, and id, or else optionally axes, and (in either case) optionally weight.

logical
Whether to return a logical value or a character string indicating the type of alluvial structure ("none", "lodes", or "alluvia")

silent
Whether to print warning messages.

key, value, id
Numeric or character; the fields of data corresponding to the axis (variable), stratum (value), and alluvium (identifying) variables.

weight
Optional numeric or character; the fields of data corresponding to alluvium or lode weights (heights when plotted).

axes
Numeric or character vector; the field(s) of data corresponding to the axi(e)s (variable(s)).

keep
A numeric or character vector indicating which variables among those passed to axes to merge into the reshapen data by id. Alternatively, a logical value indicating whether to merge all (TRUE) or none (FALSE) of these variables.

distill
A logical value indicating whether to include variables, other than those passed to key and value, that vary within values of id. Alternatively, the name of a function to be used to distill each such variable to a single value. Acceptable function names include "first", "last", and "most" (which returns the modal value; assumed if TRUE).

Details

to_lodes takes a data frame with several designated variables to be used as axes in an alluvial diagram, and reshapes the data frame so that the axis variable names constitute a new factor variable and their values comprise another. Other variables' values will be repeated, and a row-grouping variable can be introduced. This function invokes gather_. to_alluvia takes a data frame with axis and axis value variables to be used in an alluvial diagram, and reshape the data frame so that the axes constitute separate variables whose values are given by the value variable. This function invokes spread_.

Examples

# Titanic data in alluvia format
titanic_alluvia <- as.data.frame(Titanic)
head(titanic_alluvia)

#>   Class    Sex   Age Survived Freq
#> 1   1st   Male Child       No    0
#> 2   2nd   Male Child       No    0
#> 3   3rd   Male Child       No   35
#> 4  Crew   Male Child       No    0
#> 5   1st Female Child       No    0
#> 6   2nd Female Child       No    0

is_alluvial(titanic_alluvia,
            weight = "Freq")

#> [1] TRUE

# Titanic data in lodes format
titanic_lodes <- to_lodes(titanic_alluvia,
                          key = "x", value = "stratum", id = "alluvium",
                          axes = 1:4)
head(titanic_lodes)

#>   Freq alluvium     x stratum
#> 1    0        1 Class     1st
#> 2    0        2 Class     2nd
#> 3   35        3 Class     3rd
#> 4    0        4 Class    Crew
#> 5    0        5 Class     1st
#> 6    0        6 Class     2nd

is_alluvial(titanic_lodes,
            key = "x", value = "stratum", id = "alluvium",
            weight = "Freq")

#> [1] TRUE

# again in lodes format, this time keeping 'Class' as a variable
titanic_lodes2 <- to_lodes(titanic_alluvia,
                           key = "variable", value = "value", id = "passenger",
                           axes = 1:3, keep = "Class")
head(titanic_lodes2)

#>   passenger Survived Freq variable value Class
#> 1         1       No    0    Class   1st   1st
#> 2         1       No    0      Sex  Male   1st
#> 3         1       No    0      Age Child   1st
#> 4         2       No    0    Class   2nd   2nd
#> 5         2       No    0      Sex  Male   2nd
#> 6         2       No    0      Age Child   2nd

is_alluvial(titanic_lodes2,
            key = "variable", value = "value", id = "passenger",
            weight = "Freq")

#> [1] TRUE

# curriculum data in lodes format
data(majors)
head(majors)

#>   student semester curriculum
#> 1       1    CURR1   Painting
#> 2       2    CURR1   Painting
#> 3       6    CURR1   Sculpure
#> 4       8    CURR1   Painting
#> 5       9    CURR1   Sculpure
#> 6      10    CURR1   Painting

is_alluvial(majors,
            key = "semester", value = "curriculum", id = "student",
            logical = FALSE)

#> [1] "lodes"

# curriculum data in alluvia format
majors_alluvia <- to_alluvia(
  majors,
  key = "semester", value = "curriculum", id = "student"
)
head(majors_alluvia)

#>   student    CURR1       CURR3       CURR5    CURR7    CURR9   CURR11   CURR13
#> 1       1 Painting    Painting    Painting Painting Painting Painting Painting
#> 2       2 Painting    Painting    Painting Painting Painting Painting     <NA>
#> 3       6 Sculpure    Sculpure    Painting Painting Painting Painting Painting
#> 4       8 Painting    Painting    Painting Painting     <NA> Painting Painting
#> 5       9 Sculpure Art History Art History Painting Painting Painting Painting
#> 6      10 Painting    Painting    Painting Painting Painting Painting     <NA>
#>     CURR15
#> 1 Painting
#> 2     <NA>
#> 3 Painting
#> 4 Painting
#> 5 Painting
#> 6     <NA>

is_alluvial(majors_alluvia,
            axes = 2:9,
            logical = FALSE)

#> Missing alluvia for some stratum combinations.

#> [1] "alluvia"

# distill variables that vary within 'id' values
set.seed(1)
majors$hypo_grade <- LETTERS[sample(5, size = nrow(majors), replace = TRUE)]
majors_alluvia2 <- to_alluvia(
  majors,
  key = "semester", value = "curriculum", id = "student",
  distill = "most"
)

#> Distilled variables: hypo_grade

head(majors_alluvia2)

#>   student hypo_grade    CURR1       CURR3       CURR5    CURR7    CURR9
#> 1       1          B Painting    Painting    Painting Painting Painting
#> 2       2          B Painting    Painting    Painting Painting Painting
#> 3       6          C Sculpure    Sculpure    Painting Painting Painting
#> 4       8          B Painting    Painting    Painting Painting     <NA>
#> 5       9          B Sculpure Art History Art History Painting Painting
#> 6      10          B Painting    Painting    Painting Painting Painting
#>     CURR11   CURR13   CURR15
#> 1 Painting Painting Painting
#> 2 Painting     <NA>     <NA>
#> 3 Painting Painting Painting
#> 4 Painting Painting Painting
#> 5 Painting Painting Painting
#> 6 Painting     <NA>     <NA>