Deprecated functions

These functions are deprecated in the current version and will be removed for the eventual CRAN release.

ggalluvial(...)

ggalluvial_lodes(..., incl.strata = TRUE, incl.labels = FALSE)

ggalluvial_alluvia(..., incl.strata = TRUE, incl.labels = FALSE)

ggalluvial_formula(formula, data = NULL, weight, incl.strata = TRUE,
  incl.labels = TRUE, ...)

Arguments

...
Arguments passed to ggplot and inherited by geom_alluvium and geom_stratum.

incl.strata
Logical; whether to plot strata over alluvia.

incl.labels
Logical; whether to plot labels over strata.

formula
A formula to specify the axes and alluvial divisions.

data
A data frame or frequency table.

weight
A weight variable, from data or of compatible length with the elements of formula.

See also

stat_stratum, stat_flow, stat_alluvium, geom_stratum, geom_lode, geom_flow, and geom_alluvium.