ggalluvial: Alluvial Plots in 'ggplot2'

Alluvial plots use variable-width ribbons and stacked bar plots to represent multi-dimensional or repeated-measures data with categorical or ordinal variables; see Riehmann, Hanfler, and Froehlich (2005) doi:10.1109/INFVIS.2005.1532152 and Rosvall and Bergstrom (2010) doi:10.1371/journal.pone.0008694 . Alluvial plots are statistical graphics in the sense of Wilkinson (2006) doi:10.1007/0-387-28695-0 ; they share elements with Sankey diagrams and parallel sets plots but are uniquely determined from the data and a small set of parameters. This package extends Wickham's (2010) doi:10.1198/jcgs.2009.07098 layered grammar of graphics to generate alluvial plots from tidy data.

Acknowledgments

Many users identified problems and suggested improvements via email and the GitHub issue tracker.

Development benefitted from the use of equipment and the support of colleagues at UConn Health and at UF Health.

See also

Author

Maintainer: Jason Cory Brunson cornelioid@gmail.com

Authors:

Quentin D. Read