Adjoin a dataset to itself

This function binds a dataset to itself along adjacent pairs of a key variable. It is invoked by geom_flow() to convert data in lodes form to something similar to alluvia form.

self_adjoin(data, key, by = NULL, link = NULL, keep.x = NULL,
  keep.y = NULL, suffix = c(".x", ".y"))

Arguments

data
A data frame in lodes form (repeated measures data; see alluvial-data).

key
Column of data indicating sequential collection; handled as in tidyr::spread().

by
Character vector of variables to self-adjoin by; passed to dplyr::join functions.

link
Character vector of variables to adjoin. Will be replaced by pairs of variables suffixed by suffix.

keep.x, keep.y
Character vector of variables to associate with the first (respectively, second) copy of data after adjoining. These variables can overlap with each other but cannot overlap with by or link.

suffix
Suffixes to add to the adjoined link variables; passed to dplyr::join functions.

Details

self_adjoin invokes dplyr::join functions in order to convert a dataset with measures along a discrete key variable into a dataset consisting of column bindings of these measures (by any by variables) along adjacent values of key.

See also

Examples

# self-adjoin `majors` data
data(majors)
major_changes <- self_adjoin(majors, key = semester,
                             by = "student", link = c("semester", "curriculum"))
major_changes$change <- major_changes$curriculum.x == major_changes$curriculum.y
head(major_changes)

#>   step student semester.x curriculum.x semester.y curriculum.y change
#> 1    1       1      CURR1     Painting      CURR3     Painting   TRUE
#> 2    1       2      CURR1     Painting      CURR3     Painting   TRUE
#> 3    1       6      CURR1     Sculpure      CURR3     Sculpure   TRUE
#> 4    1       8      CURR1     Painting      CURR3     Painting   TRUE
#> 5    1       9      CURR1     Sculpure      CURR3  Art History  FALSE
#> 6    1      10      CURR1     Painting      CURR3     Painting   TRUE

# self-adjoin `vaccinations` data
data(vaccinations)
vaccination_steps <- self_adjoin(vaccinations, key = survey, by = "subject",
                                 link = c("survey", "response"),
                                 keep.x = c("freq", "a"))
head(vaccination_steps)

#>   step subject  survey.x response.x freq           a  survey.y response.y
#> 1    1       1 ms153_NSA     Always   48 0.050367261 ms432_NSA     Always
#> 2    1       2 ms153_NSA     Always    9 0.009443861 ms432_NSA     Always
#> 3    1       3 ms153_NSA     Always   66 0.069254984 ms432_NSA    Missing
#> 4    1       4 ms153_NSA     Always    1 0.001049318 ms432_NSA    Missing
#> 5    1       5 ms153_NSA     Always   11 0.011542497 ms432_NSA    Missing
#> 6    1       6 ms153_NSA     Always    1 0.001049318 ms432_NSA      Never

vaccination_steps <- self_adjoin(vaccinations, key = survey, by = "subject",
                                 link = c("survey", "response"),
                                 keep.x = c("freq", "a"), keep.y = "a")
head(vaccination_steps)

#>   step subject  survey.x response.x freq         a.x  survey.y response.y
#> 1    1       1 ms153_NSA     Always   48 0.050367261 ms432_NSA     Always
#> 2    1       2 ms153_NSA     Always    9 0.009443861 ms432_NSA     Always
#> 3    1       3 ms153_NSA     Always   66 0.069254984 ms432_NSA    Missing
#> 4    1       4 ms153_NSA     Always    1 0.001049318 ms432_NSA    Missing
#> 5    1       5 ms153_NSA     Always   11 0.011542497 ms432_NSA    Missing
#> 6    1       6 ms153_NSA     Always    1 0.001049318 ms432_NSA      Never
#>           a.y
#> 1 0.050367261
#> 2 0.009443861
#> 3 0.069254984
#> 4 0.001049318
#> 5 0.011542497
#> 6 0.001049318